Transfer module reload postinst script to wireguard-dkms package
The wireguard metapackage used to host the postinst script which would
reload the kernel module on upgrade.

But since the package split happened, and the wireguard-dkms package
is now built from another source package, there is no guarantee that
the cadence of the upgrade of the metapackage will coincide with the
cadence of the kernel module package.

So we move the reload script into wireguard-dkms, and we explicitly
depend on the version of wireguard-dkms that has the postinst script.

In the event where both wireguard and wireguard-dkms have this
overlapping postinst script, that should be ok because the module
reload logic should be clever enough to do the right thing as long as
the postinst scripts aren't run in parallel.